可以打开设置，顺便把GameStartOptionsViewModel分出来了
