+++ ViewModels/MainWindowViewModel.cs
@@ -8,7 +8,8 @@ using ATC4_HQ.Models; // 引入 GameModel 的命名空间
 using System.IO.Compression; // 用于解压缩功能
 using System.Text.Json;
 using master.Globals;
-using Masuit.Tools.Files; // 用于文件操作
+using Masuit.Tools.Files;
+using Newtonsoft.Json.Converters; // 用于文件操作
 
 namespace ATC4_HQ.ViewModels
 {
@@ -46,7 +47,8 @@ namespace ATC4_HQ.ViewModels
 
         private void OnSetting()
         {
-            Console.WriteLine("ViewModel: 设置逻辑。");
+            Console.WriteLine("启动设置");
+            CurrentPage = new SettingViewModel(); 
         }
 
         // 处理游戏安装和解压的通用方法，现在接收 GameModel 对象
@@ -63,35 +65,6 @@ namespace ATC4_HQ.ViewModels
             ini = new IniFile(GlobalPaths.GamePath + @"\GameData.ini");
             ini.SetValue("GameSettings" , "GameName" , GlobalPaths.GameName);
             ini.Save();
-        }
-    }
-
-    // GameStartOptionsViewModel 的定义，它必须在 ATC4_HQ.ViewModels 命名空间内
-    public partial class GameStartOptionsViewModel : ViewModelBase
-    {
-        private readonly MainWindowViewModel _mainWindowViewModel;
-
-        public ICommand Button1Command {  get; } // 启动上一次游戏
-        public ICommand Button3Command {  get; } // 启动选择游戏
-
-        // 修改构造函数：接收 MainWindowViewModel 实例
-        public GameStartOptionsViewModel(MainWindowViewModel mainWindowViewModel)
-        {
-            _mainWindowViewModel = mainWindowViewModel; // 保存引用
-            Button1Command = new RelayCommand(OnLaunchLastGame);
-            Button3Command = new RelayCommand(OnListAllGames);
-        }
-
-        private void OnLaunchLastGame()
-        {
-            Console.WriteLine("Game Start Options: 第一个按钮被点击了！尝试启动上一次游戏。");
-            
-        }
-
-        private async void OnListAllGames()
-        {
-            Console.WriteLine("Game Start Options: 第三个按钮被点击了！尝试启动选择的游戏。");
-
         }
     }
 }
