+++ ViewModels/SettingViewModel.cs
@@ -1,0 +1,8 @@
+
+
+namespace ATC4_HQ.ViewModels;
+
+public class SettingViewModel : ViewModelBase
+{
+    
+}
