+++ Views/Setting.axaml.cs
@@ -1,0 +1,13 @@
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace ATC4_HQ.Views;
+
+public partial class Setting : UserControl
+{
+    public Setting()
+    {
+        InitializeComponent();
+    }
+}
